Added Main Abstract Class and a class for each difficulty within a folder
