+++ app/src/main/java/app/plugbrain/android/di/ChallengesModule.kt
@@ -1,7 +1,9 @@
 package app.plugbrain.android.di
 
 import app.plugbrain.android.challenges.Challenge
-import app.plugbrain.android.challenges.SquareChallenge
+import app.plugbrain.android.challenges.square.SquareChallengeEasyTwoThroughFive
+import app.plugbrain.android.challenges.square.SquareChallengeMediumSixThroughTen
+import app.plugbrain.android.challenges.square.SquareChallengeHardElevenThroughTwenty
 import app.plugbrain.android.challenges.addition.AdditionFiveDigitsChallenge
 import app.plugbrain.android.challenges.addition.AdditionFourDigitsChallenge
 import app.plugbrain.android.challenges.addition.AdditionOneAndTwoDigitsChallenge
@@ -77,7 +79,9 @@ val challengesModule = module {
   factory { MultiplicationThreeByThreeDigitChallenge() } bind Challenge::class
   factory { MultiplicationThreeByFourDigitChallenge() } bind Challenge::class
   factory { MultiplicationFourByFourDigitChallenge() } bind Challenge::class
-  factory { SquareChallenge() } bind Challenge::class
+  factory { SquareChallengeEasyTwoThroughFive() } bind Challenge::class
+  factory { SquareChallengeMediumSixThroughTen() } bind Challenge::class
+  factory { SquareChallengeHardElevenThroughTwenty() } bind Challenge::class
   // New generated challenges will be added here, do not remove
 
   factory {
