+++ app/src/main/java/app/plugbrain/android/ui/challenges/compose/SquareChallengeView.kt
@@ -3,6 +3,9 @@ package app.plugbrain.android.ui.challenges.compose
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import app.plugbrain.android.challenges.SquareChallenge
+import app.plugbrain.android.challenges.square.SquareChallengeEasyTwoThroughFive
+import app.plugbrain.android.challenges.square.SquareChallengeMediumSixThroughTen
+import app.plugbrain.android.challenges.square.SquareChallengeHardElevenThroughTwenty
 
 @Composable
 fun SquareChallengeView(
